Add ability to connect to TCP server
The intention is to create a different thread to transmit the quadcopter
telemetry to the Mission Control Center (MCC) for monitoring. At the
same time this TCP connection can be used to serve as a heart beat, if
the connection is dropped, the quadcopter should immediately cut power
to motors.