Refactor PID calculation to remove derivative kick
If one implemented the derivative controller as:

    dError / dt

It will introduce something called a derivative kick, causing the output
to being a very large number very quickly. To get rid of this, the
solution is as follows:

    # when setpoint is not constant
    dError / dt = dSetpoint / dt - dInput / dt

    # when setpoint is constant
    dError / dt = - dInput / dt

This winds up being a perfect solution. Instead of adding (Kd *
derivative of Error), we subtract (Kd * derivative of Input). This is
known as using "Derivative on Measurement"

The result is that we don't end up with derivative spikes when there is
an error.

source:  http://brettbeauregard.com/blog/2011/04/improving-the-beginners-pid-introduction/